update weight datatype exp

+++ Com.Shamiraa.Service.Warehouse.Lib/Facades/ExpeditionFacade.cs
@@ -217,7 +217,7 @@ namespace Com.Shamiraa.Service.Warehouse.Lib.Facades
             {
                 try
                 {
-                    int totalweight = 0;
+                    double totalweight = 0;
                     string code = GenerateCode("ALS-KB/EXP");
                     
                     model.Code = code;
